# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiError {
  error: string;
}

/**
 * A single cell — empty, "X", or "O".
 */
export type CellValue = (typeof CellValue)[keyof typeof CellValue];

export const CellValue = {
  "": "",
  X: "X",
  O: "O",
} as const;

/**
 * Which player the AI is playing as.
 */
export type Player = (typeof Player)[keyof typeof Player];

export const Player = {
  X: "X",
  O: "O",
} as const;

/**
 * How strong the AI should play. The agent blends its heuristic move
with a uniformly-random move using this probability of playing the
heuristic (optimal) move:
  easy   = 0.10  (≈ 90% random — barely above random)
  medium = 0.525 (exact midpoint between easy and hard)
  hard   = 0.95  (≈ 5% random — barely below perfect)

 */
export type AiMoveRequestDifficulty =
  (typeof AiMoveRequestDifficulty)[keyof typeof AiMoveRequestDifficulty];

export const AiMoveRequestDifficulty = {
  easy: "easy",
  medium: "medium",
  hard: "hard",
} as const;

export interface AiMoveRequest {
  /**
   * @minimum 2
   * @maximum 10
   */
  rows: number;
  /**
   * @minimum 2
   * @maximum 10
   */
  cols: number;
  /**
   * Depth (z) of the board. 1 = 2D game, >1 = 3D game.

   * @minimum 1
   * @maximum 10
   */
  depth?: number;
  /**
   * Size along the 4th dimension (w). 1 = 2D/3D game, >1 = 4D game.

   * @minimum 1
   * @maximum 4
   */
  w?: number;
  /**
   * Number of consecutive matching markers required to win.
Forwarded to the Python agent so it can recognize wins/blocks
for the active game's win-length setting.

   * @minimum 2
   * @maximum 10
   */
  winLen?: number;
  /** How strong the AI should play. The agent blends its heuristic move
with a uniformly-random move using this probability of playing the
heuristic (optimal) move:
  easy   = 0.10  (≈ 90% random — barely above random)
  medium = 0.525 (exact midpoint between easy and hard)
  hard   = 0.95  (≈ 5% random — barely below perfect)
 */
  difficulty?: AiMoveRequestDifficulty;
  /** Flattened board of length rows*cols*depth*w.
Index = w*(rows*cols*depth) + z*(rows*cols) + y*cols + x,
where y is row, x is column, z is layer, w is hyperslice.
 */
  board: CellValue[];
  player: Player;
}

export interface AiMoveResponse {
  /** @minimum 0 */
  row: number;
  /** @minimum 0 */
  col: number;
  /**
   * Layer index (z). Omit or 0 for 2D games.
   * @minimum 0
   */
  z?: number;
  /**
   * Hyperslice index (w). Omit or 0 for 2D/3D games.
   * @minimum 0
   */
  w?: number;
}